+++ R/server.R
@@ -67,8 +67,9 @@
 #' * `mcp_server()` blocks the R process it's called in indefinitely and isn't
 #'   intended for interactive use.
 #' * `mcp_session()` makes the interactive R session it's called in available to
-#'   MCP servers. It returns the \pkg{nanonext} socket used for communicating
-#'   with the server. Call [close()] on the socket to stop the session.
+#'   MCP servers. It returns invisibly the \pkg{nanonext} socket used for
+#'   communicating with the server. Call [close()] on the socket to stop the
+#'   session.
 #'
 #' @seealso
 #' - The "R as an MCP server" vignette at